Updating to Qt 6 (#523)
* Remove some deprecated calls.

* Update CMake Lists

* Update OIGTL/OIGTLIO

* VTK fix

* The macro SendMessage in winuser.h clashes with the method name in igtlioConnector. I would guess Qt6 now includes this somewhere where Qt5 wasn't. The cleaner way to resolve this would be to rename the SendMessage method in OIGTLIO to something unique, but this would require also changing every other application using OIGTLIO.

* Update Windows and Linux CI to build with Qt6

+++ IbisLib/worldobject.cpp
@@ -74,7 +74,7 @@ void WorldObject::SetCursorColor( const QColor & c )
     this->GetManager()->SetCursorColor( c );
 }
 
-void WorldObject::SetCursorColor( double color[3] )
+void WorldObject::SetCursorColor( float color[3] )
 {
     Q_ASSERT( this->GetManager() );
     QColor col( (int)( color[0] * 255 ), (int)( color[1] * 255 ), (int)( color[2] * 255 ) );
